MacPUF: A Simulation of Human Respiration, Gas Exchange and Control
This project is an adaptation of MacPUF, originally published in 1977 by C.J. Dickinson.  This was a FORTRAN program used by medical students and physiology researchers to simulate human respiration, gas exchange and control on a time sharing computer system.  I translated this program into BASIC in 1980, and subsequently had the ambition to translate it into C to make it able to run on early Macintosh computers.  Alas, for time reasons, this project has been delayed until the present, when I have successfully ported the program to run on iOS platforms using SwiftUI.